# apps/shop-admin-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/locations": {
      "post": {
        "operationId": "LocationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Locations"
        ]
      },
      "get": {
        "operationId": "LocationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Locations"
        ]
      }
    },
    "/locations/{id}": {
      "patch": {
        "operationId": "LocationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLocationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Locations"
        ]
      }
    },
...
    "/orders": {
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedOrders"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shop",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateLocationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "accountId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "addressCity": {
            "type": "string",
            "nullable": true
          },
          "addressState": {
            "type": "string",
            "nullable": true
          },
          "addressPostalCode": {
            "type": "string",
            "nullable": true
          },
          "addressCountry": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "accountId",
          "name",
          "addressLine1",
          "addressLine2",
          "addressCity",
          "addressState",
          "addressPostalCode",
          "addressCountry",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateLocationDto": {
        "type": "object",
        "properties": {
          "addressLine1": {
            "type": "string",
            "nullable": true
          },
          "addressLine2": {
            "type": "string",
            "nullable": true
          },
          "addressCity": {
            "type": "string",
            "nullable": true
          },
          "addressState": {
            "type": "string",
            "nullable": true
          },
          "addressPostalCode": {
            "type": "string",
            "nullable": true
          },
          "addressCountry": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "OrderListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "customerName": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "itemCount": {
            "type": "number"
          },
          "amountTotalCents": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerName",
          "customerEmail",
          "itemCount",
          "amountTotalCents",
          "createdAt"
        ]
      },
      "PaginatedOrders": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderListItem"
            }
          },
          "total": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ]
      },
      "OrderDetailItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "variantId": {
            "type": "number",
            "nullable": true
          },
          "productName": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "optionsLabel": {
            "type": "string",
            "nullable": true
          },
          "priceCents": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "variantId",
          "productName",
          "sku",
          "optionsLabel",
          "priceCents",
          "quantity"
        ]
      },
      "OrderDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "customerName": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "shippingLine1": {
            "type": "string"
          },
          "shippingLine2": {
            "type": "string",
            "nullable": true
          },
          "shippingCity": {
            "type": "string"
          },
          "shippingState": {
            "type": "string",
            "nullable": true
          },
          "shippingPostalCode": {
            "type": "string"
          },
          "shippingCountry": {
            "type": "string"
          },
          "subtotalCents": {
            "type": "number"
          },
          "amountTotalCents": {
            "type": "number"
          },
          "shippingCents": {
            "type": "number"
          },
          "shippingLocationId": {
            "type": "number",
            "nullable": true
          },
          "shippingCarrier": {
            "type": "string",
            "nullable": true
          },
          "shippingServiceLevel": {
            "type": "string",
            "nullable": true
          },
          "trackingNumber": {
            "type": "string",
            "nullable": true
          },
          "trackingUrl": {
            "type": "string",
            "nullable": true
          },
          "labelUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderDetailItem"
            }
          }
        },
        "required": [
          "id",
          "customerName",
          "customerEmail",
          "shippingLine1",
          "shippingLine2",
          "shippingCity",
          "shippingState",
          "shippingPostalCode",
          "shippingCountry",
          "subtotalCents",
          "amountTotalCents",
          "shippingCents",
          "shippingLocationId",
          "shippingCarrier",
          "shippingServiceLevel",
          "trackingNumber",
          "trackingUrl",
          "labelUrl",
          "createdAt",
          "items"
        ]
      },
...
    }
  }
}